requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2
argon2-cffi
bcrypt
pytest
httpx